% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covfam.R
\name{covfam}
\alias{covfam}
\title{CovBat Family Harmonization}
\usage{
covfam(
  data,
  covar,
  bat,
  model,
  formula,
  score.model = NULL,
  score.args = NULL,
  eb = TRUE,
  robust.LS = FALSE,
  percent.var = 0.95,
  n.pc = NULL,
  std.var = TRUE,
  debug = FALSE,
  ...
)
}
\arguments{
\item{data}{\emph{n x p} data frame or matrix of observations where
\emph{p} is the number of features and \emph{n} is the number of subjects.}

\item{covar}{Data frame or matrix of covariates supplied to `model`}

\item{bat}{Factor indicating batch (often equivalent to site or scanner)}

\item{model}{Model function. ComBat Family supports any models that take
arguments `formula` and `data`, but are limited to models fitting with
identity link (e.g. `family = gaussian(link = "identity")`). This includes
\link[stats]{lm}, \link[mgcv]{gam}, \link[gamlss]{gamlss},
\link[quantreg]{rq}, \link[lme4]{lmer}, and more}

\item{formula}{Formula for `model`, format is dependent on choice of model.}

\item{score.model}{Model for scores, defaults to NULL for fitting basic
location and scale model without covariates on the scores}

\item{score.args}{List of arguments for score model, requires `formula`}

\item{eb}{If \code{TRUE}, uses ComBat model with empirical Bayes for mean
and variance harmonization.}

\item{robust.LS}{If \code{TRUE}, uses robust location and scale estimators
for error variance and site effect parameters. Currently uses median and
biweight midvariance}

\item{percent.var}{Numeric. The number of harmonized principal component
scores is selected to explain this proportion of the variance.}

\item{n.pc}{Optional numeric. If specified, this number of principal
component scores is harmonized. Overrides \code{percent.var}.}

\item{std.var}{If \code{TRUE}, scales variances to be equal to 1 before PCA.}

\item{debug}{Whether to output model fits and intermediate data frames}

\item{...}{Additional arguments to `model`}
}
\description{
Implementation of the CovBat Family of harmonization methods allowing for
removal of multivariate batch effects, flexible covariate modeling and
alternative estimators for site effect adjustment. Support for modeling of
both location and scale via GAMLSS. Additional support for modeling of
covariate effects in score location and scale.
}
