% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comfam.R
\name{deidentify.comfam}
\alias{deidentify.comfam}
\title{Deidentify ComBat Family Harmonization Fit}
\usage{
\method{deidentify}{comfam}(object)
}
\arguments{
\item{object}{Object of class \code{comfam}, typically output of
\link[ComBatFamily]{comfam}}
}
\value{
\code{deidentify.comfam} returns a list containing the following components:
\item{batch.info}{New batch information, including reference batch if specified}
\item{fits}{Deidentified list of model fits from regression step, forwarded from \code{object}}
\item{estimates}{List of estimates from standardization and batch effect correction, including new batches if relevant}
}
\description{
Remove original data from \code{comfam} fit. The deidentified harmonization fit
can then be shared with other sites for use in
\link[ComBatFamily]{predict.comfam}.
}
\examples{
com_out <- comfam(iris[1:75,1:2], iris$Species[1:75], iris[1:75,3:4], lm, y ~ Petal.Length + Petal.Width)
com_deid <- deidentify(com_out)

predict(com_deid, iris[76:150,1:2], iris$Species[76:150], iris[1:75,3:4])
}
