% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comfam.R
\name{plot.comfam}
\alias{plot.comfam}
\title{Plot Diagnostics for \code{comfam}}
\usage{
\method{plot}{comfam}(object, feature)
}
\arguments{
\item{object}{Object of class \code{comfam}, typically output of
\link[ComBatFamily]{comfam}}

\item{feature}{Feature to diagnose, either index or variable name}
}
\value{
\code{plot.comfam} displays plots from the model fit.
}
\description{
Diagnostic plots for original model fits in \code{comfam}, leverages S3 plot
methods for \code{model} (e.g. \link[stats]{plot.lm})
}
\examples{
com_out <- comfam(iris[,1:2], iris$Species)
plot(com_out, "Sepal.Width")
}
