% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covfam.R
\name{predict.covfam}
\alias{predict.covfam}
\title{Apply CovBat Harmonization to New Data}
\usage{
\method{predict}{covfam}(
  object,
  newdata,
  newbat,
  newcovar = NULL,
  robust.LS = FALSE,
  eb = TRUE,
  score.args = NULL,
  ...
)
}
\arguments{
\item{object}{Object of class \code{covfam}, typically output of
\link[ComBatFamily]{covfam}}

\item{newdata}{\emph{n x p} data frame or matrix of new observations where
\emph{p} is the number of features and \emph{n} is the number of subjects.
The features must match the original \code{data} used in \code{object}}

\item{newbat}{Factor indicating new batch (often equivalent to site or scanner)}

\item{newcovar}{Data frame or matrix of new covariates supplied to \code{model}.
Must contain all variables specified in the original \code{formula} used in
\code{object}.}

\item{robust.LS}{If \code{TRUE}, uses robust location and scale estimators
for new batch effect estimates. Currently uses median and biweight
midvariance}

\item{eb}{If \code{TRUE}, uses CovBat model with empirical Bayes for new batches}

\item{score.args}{List of arguments to \code{predict} for the class of \code{score.model}}

\item{...}{Additional arguments to \code{predict} for the class of \code{model} (e.g.
\code{predict.lm} for CovBat)}
}
\value{
\code{predict.covfam} returns a list containing the following components:
\item{batch.info}{Batch information, including reference batch if specified}
\item{combat.out}{List output of \link[ComBatFamily]{comfam} from the ComBat step}
\item{pc.output}{Output of \code{prcomp} from PCA step}
\item{n.pc}{Numeric, number of PCs harmonized}
\item{scores.com}{List output of \link[ComBatFamily]{comfam} from the CovBat step}
}
\description{
Using parameters estimated via \code{covfam}, apply harmonization on new data.
\code{predict.covfam} will estimate new batch adjustments if new batches are
specified. For batches with existing estimates, the estimates from \code{object}
are used. Harmonization targets are the same as \code{object} (e.g. \code{ref.batch}
from \code{object} if specified). Model specifications are defined by the
original \code{covfam} fit.
}
\examples{
cov_out <- covfam(iris[1:75,1:4], iris$Species[1:75])

# out-of-sample with new batch
out_pred <- predict(cov_out, iris[76:150,1:4], iris$Species[76:150])

# in-sample
in_pred <- predict(cov_out, iris[1:25,1:4], iris$Species[1:25])
max(in_pred$dat.covbat - cov_out$dat.covbat[1:25,1:4])
}
