% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comfam.R
\name{comfam}
\alias{comfam}
\title{ComBat Family Harmonization}
\usage{
comfam(
  data,
  covar,
  bat,
  model,
  formula,
  eb = TRUE,
  robust.LS = FALSE,
  debug = FALSE,
  ...
)
}
\arguments{
\item{data}{\emph{n x p} data frame or matrix of observations where
\emph{p} is the number of features and \emph{n} is the number of subjects.}

\item{covar}{Data frame or matrix of covariates supplied to `model`}

\item{bat}{Factor indicating batch (often equivalent to site or scanner)}

\item{model}{Model function. ComBat Family supports any models that take
arguments `formula` and `data`, but are limited to models fitting with
identity link (e.g. `family = gaussian(link = "identity")`). This includes
\link[stats]{lm}, \link[mgcv]{gam}, \link[gamlss]{gamlss},
\link[quantreg]{rq}, \link[lme4]{lmer}, and more}

\item{formula}{Formula for `model`, format is dependent on choice of model.}

\item{eb}{If \code{TRUE}, uses ComBat model with empirical Bayes for mean
and variance harmonization}

\item{robust.LS}{If \code{TRUE}, uses robust location and scale estimators
for error variance and site effect parameters. Currently uses median and
biweight midvariance}

\item{debug}{Whether to output model fits and intermediate data frames}

\item{...}{Additional arguments to `model`}
}
\description{
Implementation of the ComBat Family of harmonization methods allowing for
flexible covariate modeling and alternative estimators for site effect
adjustment. Support for modeling of both location and scale via GAMLSS and
longitudinal harmonization via mixed effects models.
}
